+++ Assets/Editor/ReaderGenerator.cs
@@ -103,7 +103,6 @@ public class ReaderGenerator {
 	};
 
 	private static readonly List<string> m_complex = new List<string>() {
-		"void LogError(Exception error)",
 		"void LogError(string error)",
 		"Tater GetTater()",
 		"bool OptionalObject<T>(string name, ref T value) where T : IReadable, new()",