+++ Assets/PoDato/Scripts/IReader.cs
@@ -5,8 +5,6 @@ using UnityEngine;
 
 namespace PoDato {
 	public interface IReader {
-		void LogError(Exception error);
-		void LogError(string error);
 		Tater GetTater();
 		bool OptionalObject<T>(string name, ref T value) where T : IReadable, new();
 		bool OptionalObjectList<T, U>(string name, ref T value) where T : ICollection<U>, new () where U : IReadable, new ();