+++ Assets/PoDato/Scripts/Internal/Exceptions.cs
@@ -3,9 +3,18 @@
 namespace PoDato {
 
 	internal class ParseException : Exception {
-		internal ParseException(FilePosition pos, string message) : base($"pos{pos}: {message}") { 
-		
+		public FilePosition Position { get; }
+		public ParseException(FilePosition position, string message) : base(message) {
+			Position = position;
 		}
+	}
+	internal class DeserializationException : Exception {
+
+		public Tater Tater { get; }		
+		public DeserializationException(Tater tater, string message) : base(message) {
+			Tater = tater;
+		}
+
 	}
 
 }