+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] 2022-11-03
+### Added
+- Improved error handling for reading `Tater` objects
+- `ReadError` class containing more information about read errors than just the message
+
 ## [1.6.0] 2022-11-01
 ### Added
 - `LineNumber` property to `Tater` objects to help refer to them during reading